Random Forest Regressor
With the predicted price having a mean of 8,585.63 and a standard deviation of 2,182.46 for the houses, the estimated predicted price ranged from 2,750.30 to 14,137.30.
The predicted prices appeared to be normally distributed with 8,556.80 as the median. The skewness of the predicted price is approximately -0.0174 which corroborates that the distribution of predicted prices is approximately symmetrical meaning that the distribution has a similar number of values on either side of the mean i.e., the data is not significantly skewed to the left or right.
For our model, Random Forest Regressor was used and while with a Mean Squared Error MSE of about 80814.003, which imply that the average squared difference between the predicted prices and the actual prices is around 80814, the R-Squared of this model is approximately 0.9824 meaning that 98.24% of the variability in the predicted prices can be explained by the Random Forest Regressor model indicating that this model is a good fit for predicting prices from the dataset and that the independent variables (features) are strong predictors of the dependent variable(prices).
Conclusively, the Random Forest Regressor model used is a good model for predicting the prices of houses based on the given features.
The predicted house price values, “predicted_prices” column, from my code are exported to the attached csv file titled “price_prediction_sheet”. Please see the appendix titled Attachment 1: “Code Output Sheet”.
Comparing the summary statistics of the "price" and "predicted_price", mean of the "price" column (approximately 8601.86) is somewhat higher than the mean of the "predicted_price" (8585.63). However, the difference is quite negligible, and this could be due to random variation.
Consequentially, the standard deviation of the price column is also somewhat higher than the standard deviation of the predicted price but again, the difference is very small.
While the range of both columns is the same and the quartiles (25th, 50th, and 75th percentiles) for both columns are very similar, with only slight differences in some cases.
Generally, with this comparison, the "predicted_prices" column is seemingly a good predictor of the "price" column, as the two columns have similar summary statistics.
However, further analysis was carried out (e.g., hypothesis tests) to fully assess the accuracy of the model.
After conducting this hypothesis test, the t-stat of about 0.16 was gotten which implies that the difference between the mean price and the mean predicted price is not significantly different from zero.
Additionally, since the p-value of 0.88 is greater than the 0.05 threshold, this suggests that the Random Forest Regressor model's predictions are not significantly different from the actual prices